Use :float instead of :decimal for player stats

+++ lib/fungus_toast/players/player.ex
@@ -32,24 +32,24 @@ defmodule FungusToast.Players.Player do
 
     field :mutation_points, :integer, default: 0, null: false
 
-    field :top_left_growth_chance, :decimal, default: 0.0, null: false
-    field :top_growth_chance, :decimal, default: 0.0, null: false
-    field :top_right_growth_chance, :decimal, default: 0.0, null: false
-    field :right_growth_chance, :decimal, default: 0.0, null: false
-    field :bottom_right_growth_chance, :decimal, default: 0.0, null: false
-    field :bottom_growth_chance, :decimal, default: 0.0, null: false
-    field :bottom_left_growth_chance, :decimal, default: 0.0, null: false
-    field :left_growth_chance, :decimal, default: 0.0, null: false
+    field :top_left_growth_chance, :float, default: 0.0, null: false
+    field :top_growth_chance, :float, default: 0.0, null: false
+    field :top_right_growth_chance, :float, default: 0.0, null: false
+    field :right_growth_chance, :float, default: 0.0, null: false
+    field :bottom_right_growth_chance, :float, default: 0.0, null: false
+    field :bottom_growth_chance, :float, default: 0.0, null: false
+    field :bottom_left_growth_chance, :float, default: 0.0, null: false
+    field :left_growth_chance, :float, default: 0.0, null: false
 
     field :dead_cells, :integer, default: 0, null: false
     field :live_cells, :integer, default: 0, null: false
     field :regenerated_cells, :integer, default: 0, null: false
 
-    field :apoptosis_chance, :decimal, default: 0.0, null: false
-    field :starved_cell_death_chance, :decimal, default: 0.0, null: false
-    field :mutation_chance, :decimal, default: 0.0, null: false
-    field :regeneration_chance, :decimal, default: 0.0, null: false
-    field :mycotoxin_fungicide_chance, :decimal, default: 0.0, null: false
+    field :apoptosis_chance, :float, default: 0.0, null: false
+    field :starved_cell_death_chance, :float, default: 0.0, null: false
+    field :mutation_chance, :float, default: 0.0, null: false
+    field :regeneration_chance, :float, default: 0.0, null: false
+    field :mycotoxin_fungicide_chance, :float, default: 0.0, null: false
 
     belongs_to :user, FungusToast.Accounts.User
     belongs_to :game, FungusToast.Games.Game
